backend/requirements.txt
Flask==2.3.3
Flask-SocketIO==5.3.6
Flask-Limiter==3.6.0
Flask-Cors==4.0.1
PyJWT==2.8.0
python-dotenv==1.0.1
gunicorn==22.0.0
uvicorn==0.23.2
requests==2.31.0
redis==5.0.7
eventlet==0.33.3
# Database & ORM
Flask-SQLAlchemy==2.5.1
SQLAlchemy==1.4.51
SQLAlchemy-Utils==0.38.3
psycopg2==2.9.9
Werkzeug==2.3.8
cryptography==42.0.8
Flask-Migrate==4.0.7
Flask-Script==2.0.6
alembic==1.13.2
blinker==1.8.2
click==8.1.7
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5 